feat: 테마 삭제 시 예약 현황 확인 기능 구현
- 예약건이 남아 있을 시 테마 제거 불가

+++ src/main/java/kakao/service/ThemeService.java
@@ -3,11 +3,13 @@ package kakao.service;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import kakao.domain.Reservation;
 import kakao.domain.Theme;
 import kakao.dto.request.CreateThemeRequest;
 import kakao.dto.response.ThemeResponse;
 import kakao.error.ErrorCode;
 import kakao.error.exception.RoomReservationException;
+import kakao.repository.reservation.ReservationRepository;
 import kakao.repository.theme.ThemeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,6 +21,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+
+    private final ReservationRepository reservationRepository;
 
     public ThemeResponse createTheme(CreateThemeRequest request) {
         Theme theme = Theme.builder()
@@ -44,6 +48,14 @@ public class ThemeService {
     }
 
     public int deleteTheme(Long id) {
+        Theme theme = themeRepository.findById(id);
+        if (Objects.isNull(theme)) {
+            return 0;
+        }
+        List<Reservation> reservations = reservationRepository.findAllByTheme(theme);
+        if (reservations.size() > 0) {
+            throw new RoomReservationException(ErrorCode.THEME_RESERVATIONS_YET_EXIST);
+        }
         return themeRepository.delete(id);
     }
 }
